Added a makefile
Added a makefile to facilitate compilation, and required C++11 due to usage of "nullptr" in svmu.cpp.  The nullptr keyword was only introduced in C++09, but not incorporated into C++0x, unfortunately: http://stackoverflow.com/questions/1282295/what-exactly-is-nullptr
Tested out on our cluster (had to use G++ > 4.4.7 to compile with --std=c++11), and it compiles fine.